Replace clipping plane handling in pan,zoom and rotate by events.
