+++ packages/tools/src/enums/Events.ts
@@ -39,6 +39,12 @@ enum Events {
 
   VOLUMECROPPING_TOOL_CHANGED = 'CORNERSTONE_TOOLS_VOLUMECROPPING_TOOL_CHANGED',
 
+  ZOOM_TOOL_CHANGED = 'CORNERSTONE_TOOLS_ZOOM_TOOL_CHANGED',
+
+  PAN_TOOL_CHANGED = 'CORNERSTONE_TOOLS_PAN_TOOL_CHANGED',
+
+  TRACKBALLROTATE_TOOL_CHANGED = 'CORNERSTONE_TOOLS_TRACKBALLROTATE_TOOL_CHANGED',
+
   ///////////////////////////////////////
   //            Annotations
   ///////////////////////////////////////
