+++ packages/tools/src/tools/PanTool.ts
@@ -1,9 +1,14 @@
 import { BaseTool } from './base';
-import { getEnabledElement } from '@cornerstonejs/core';
+import {
+  getEnabledElement,
+  triggerEvent,
+  eventTarget,
+} from '@cornerstonejs/core';
+import type { EventTypes, PublicToolProps, ToolProps } from '../types';
+import { Events } from '../enums';
 import type { Types } from '@cornerstonejs/core';
 import vtkPlane from '@kitware/vtk.js/Common/DataModel/Plane';
 
-import type { EventTypes, PublicToolProps, ToolProps } from '../types';
 import { getToolGroup } from '../store/ToolGroupManager';
 /**
  * Tool that pans the camera in the plane defined by the viewPlaneNormal and the viewUp.
@@ -25,74 +30,6 @@ class PanTool extends BaseTool {
 
   mouseDragCallback(evt: EventTypes.InteractionEventType) {
     this._dragCallback(evt);
-  }
-  _transformNormal(normal: Types.Point3, mat: number[]): Types.Point3 {
-    return [
-      mat[0] * normal[0] + mat[3] * normal[1] + mat[6] * normal[2],
-      mat[1] * normal[0] + mat[4] * normal[1] + mat[7] * normal[2],
-      mat[2] * normal[0] + mat[5] * normal[1] + mat[8] * normal[2],
-    ];
-  }
-
-  _updateClippingPlanes(viewport) {
-    const actorEntry = viewport.getDefaultActor();
-    const actor = actorEntry.actor;
-    const mapper = actor.getMapper();
-    const matrix = actor.getMatrix();
-
-    // Extract rotation part for normals
-    const rot = [
-      matrix[0],
-      matrix[1],
-      matrix[2],
-      matrix[4],
-      matrix[5],
-      matrix[6],
-      matrix[8],
-      matrix[9],
-      matrix[10],
-    ];
-
-    // --- Get the VolumeCroppingTool instance for this viewport ---
-    const toolGroup = getToolGroup(this.toolGroupId);
-    const croppingTool = toolGroup?.getToolInstance?.('VolumeCroppingTool');
-    // Use the tool's originalClippingPlanes property
-    const originalPlanes = croppingTool?.originalClippingPlanes;
-    if (!originalPlanes || !originalPlanes.length) {
-      return;
-    }
-
-    mapper.removeAllClippingPlanes();
-    originalPlanes.forEach((plane) => {
-      const origin =
-        typeof plane.getOrigin === 'function'
-          ? plane.getOrigin()
-          : plane.origin;
-      const normal =
-        typeof plane.getNormal === 'function'
-          ? plane.getNormal()
-          : plane.normal;
-
-      // Transform origin (full 4x4)
-      const o: Types.Point3 = [
-        matrix[0] * origin[0] +
-          matrix[4] * origin[1] +
-          matrix[8] * origin[2] +
-          matrix[12],
-        matrix[1] * origin[0] +
-          matrix[5] * origin[1] +
-          matrix[9] * origin[2] +
-          matrix[13],
-        matrix[2] * origin[0] +
-          matrix[6] * origin[1] +
-          matrix[10] * origin[2] +
-          matrix[14],
-      ];
-      // Transform normal (rotation only)
-      const n: Types.Point3 = this._transformNormal(normal, rot);
-      const planeInstance = vtkPlane.newInstance({ origin: o, normal: n });
-      mapper.addClippingPlane(planeInstance);
-    });
   }
 
   _dragCallback(evt: EventTypes.InteractionEventType) {
@@ -127,8 +64,9 @@ class PanTool extends BaseTool {
       focalPoint: updatedFocalPoint,
       position: updatedPosition,
     });
-    // Update clipping planes after pan
-    this._updateClippingPlanes(enabledElement.viewport);
+    triggerEvent(eventTarget, Events.PAN_TOOL_CHANGED, {
+      viewport: enabledElement.viewport,
+    });
     enabledElement.viewport.render();
   }
 }
