+++ packages/tools/src/tools/TrackballRotateTool.ts
@@ -3,6 +3,7 @@ import vtkPlane from '@kitware/vtk.js/Common/DataModel/Plane';
 import { Events } from '../enums';
 import {
   eventTarget,
+  triggerEvent,
   getEnabledElement,
   getEnabledElementByIds,
   Enums,
@@ -154,76 +155,6 @@ class TrackballRotateTool extends BaseTool {
     }
   };
 
-  // Helper to transform a normal by a 3x3 matrix
-  _transformNormal(normal: Types.Point3, mat: number[]): Types.Point3 {
-    return [
-      mat[0] * normal[0] + mat[3] * normal[1] + mat[6] * normal[2],
-      mat[1] * normal[0] + mat[4] * normal[1] + mat[7] * normal[2],
-      mat[2] * normal[0] + mat[5] * normal[1] + mat[8] * normal[2],
-    ];
-  }
-
-  // Update all clipping planes after rotation
-  _updateClippingPlanes(viewport) {
-    const actorEntry = viewport.getDefaultActor();
-    const actor = actorEntry.actor as Types.VolumeActor;
-    const mapper = actor.getMapper();
-    const matrix = actor.getMatrix();
-    // Extract rotation part for normals
-    const rot = [
-      matrix[0],
-      matrix[1],
-      matrix[2],
-      matrix[4],
-      matrix[5],
-      matrix[6],
-      matrix[8],
-      matrix[9],
-      matrix[10],
-    ];
-
-    // --- Get the VolumeCroppingTool instance for this viewport ---
-    const toolGroup = getToolGroup(this.toolGroupId);
-    const croppingTool = toolGroup?.getToolInstance?.('VolumeCroppingTool');
-    // Use the tool's originalClippingPlanes property
-    const originalPlanes = croppingTool?.originalClippingPlanes;
-    if (!originalPlanes || originalPlanes.length === 0) {
-      return;
-    }
-
-    mapper.removeAllClippingPlanes();
-    originalPlanes.forEach((plane) => {
-      const origin =
-        typeof plane.getOrigin === 'function'
-          ? plane.getOrigin()
-          : plane.origin;
-      const normal =
-        typeof plane.getNormal === 'function'
-          ? plane.getNormal()
-          : plane.normal;
-
-      // Transform origin (full 4x4)
-      const o: Types.Point3 = [
-        matrix[0] * origin[0] +
-          matrix[4] * origin[1] +
-          matrix[8] * origin[2] +
-          matrix[12],
-        matrix[1] * origin[0] +
-          matrix[5] * origin[1] +
-          matrix[9] * origin[2] +
-          matrix[13],
-        matrix[2] * origin[0] +
-          matrix[6] * origin[1] +
-          matrix[10] * origin[2] +
-          matrix[14],
-      ];
-      // Transform normal (rotation only)
-      const n: Types.Point3 = this._transformNormal(normal, rot);
-      const planeInstance = vtkPlane.newInstance({ origin: o, normal: n });
-      mapper.addClippingPlane(planeInstance);
-    });
-  }
-
   rotateCamera = (viewport, centerWorld, axis, angle) => {
     const vtkCamera = viewport.getVtkActiveCamera();
     const viewUp = vtkCamera.getViewUp();
@@ -255,8 +186,9 @@ class TrackballRotateTool extends BaseTool {
       focalPoint: newFocalPoint,
     });
 
-    // Update clipping planes after rotation
-    this._updateClippingPlanes(viewport);
+    triggerEvent(eventTarget, Events.PAN_TOOL_CHANGED, {
+      viewport: viewport,
+    });
   };
 
   _dragCallback(evt: EventTypes.InteractionEventType): void {
