+++ packages/tools/src/tools/VolumeCroppingTool.ts
@@ -551,7 +551,81 @@ class VolumeCroppingTool extends AnnotationTool {
         this._onControlToolChange(evt);
       }
     );
+    eventTarget.addEventListener(Events.ZOOM_TOOL_CHANGED, (evt) => {
+      this._updateClippingPlanes(evt.detail.viewport);
+    });
+    eventTarget.addEventListener(Events.PAN_TOOL_CHANGED, (evt) => {
+      this._updateClippingPlanes(evt.detail.viewport);
+    });
+    eventTarget.addEventListener(Events.TRACKBALLROTATE_TOOL_CHANGED, (evt) => {
+      this._updateClippingPlanes(evt.detail.viewport);
+    });
   };
+
+  _transformNormal(normal: Types.Point3, mat: number[]): Types.Point3 {
+    return [
+      mat[0] * normal[0] + mat[3] * normal[1] + mat[6] * normal[2],
+      mat[1] * normal[0] + mat[4] * normal[1] + mat[7] * normal[2],
+      mat[2] * normal[0] + mat[5] * normal[1] + mat[8] * normal[2],
+    ];
+  }
+
+  _updateClippingPlanes(viewport) {
+    // Get the actor and transformation matrix
+    const actorEntry = viewport.getDefaultActor();
+    const actor = actorEntry.actor;
+    const mapper = actor.getMapper();
+    const matrix = actor.getMatrix();
+
+    // Extract rotation part for normals
+    const rot = [
+      matrix[0],
+      matrix[1],
+      matrix[2],
+      matrix[4],
+      matrix[5],
+      matrix[6],
+      matrix[8],
+      matrix[9],
+      matrix[10],
+    ];
+
+    const originalPlanes = this.originalClippingPlanes;
+    if (!originalPlanes || !originalPlanes.length) {
+      return;
+    }
+
+    mapper.removeAllClippingPlanes();
+    originalPlanes.forEach((plane) => {
+      const origin =
+        typeof plane.getOrigin === 'function'
+          ? plane.getOrigin()
+          : plane.origin;
+      const normal =
+        typeof plane.getNormal === 'function'
+          ? plane.getNormal()
+          : plane.normal;
+      // Transform origin (full 4x4)
+      const o: Types.Point3 = [
+        matrix[0] * origin[0] +
+          matrix[4] * origin[1] +
+          matrix[8] * origin[2] +
+          matrix[12],
+        matrix[1] * origin[0] +
+          matrix[5] * origin[1] +
+          matrix[9] * origin[2] +
+          matrix[13],
+        matrix[2] * origin[0] +
+          matrix[6] * origin[1] +
+          matrix[10] * origin[2] +
+          matrix[14],
+      ];
+      // Transform normal (rotation only)
+      const n: Types.Point3 = this._transformNormal(normal, rot);
+      const planeInstance = vtkPlane.newInstance({ origin: o, normal: n });
+      mapper.addClippingPlane(planeInstance);
+    });
+  }
 
   _onControlToolChange = (evt) => {
     // coronal  is y axis in green
